pyproject.toml
[project]
name = "running-contacts"
version = "0.1.0"
description = "Outil local pour croiser contacts et résultats de courses"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "rapidfuzz",
    "google-api-python-client",
    "google-auth-oauthlib",
]

[project.scripts]
running-contacts = "running_contacts.cli:app"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
